[TASK] Rename datatables.js to be version independant
Unline jQuery which has several versions included
in TYPO3 and an API for loading multiple versions,
the core should ship jQuery.dataTables in only
one version. In order to make minor upgrades easy
for this contrib file, the JS file that was added during
the Hamburg UX sprint,  should be renamed to be
version-independant. This way, extensions can include
the shipped core version without having to be
changed all the time once the plugin is updated in
the core.

This is the same behaviour as for all third-party
JS libraries except ExtJS and jQuery core.

Releases: master
Resolves: #63435
Change-Id: I81fb9883940db45e148e127a8a60e2f67aefe40b
Reviewed-on: http://review.typo3.org/34786
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>