[!!!][FEATURE] Flash message view helper add deferred rendering
Deprecate renderMode in favor of a flexible deferred
rendering of flash messages in the Fluid template.

This means that flash messages should no longer contain HTML tags
but only plain text, if rendered by the default rendering.

The rendered output will by default now be a <ul> list.
The usage of renderMode in the Core is removed and styles are
adapted to match the previous output.

The core is adapted to now use htmlspecialchars in a controller context
for user input. Encoding is now taken care of during rendering of
the Fluid template.

Resolves: #63453
Releases: master
Change-Id: Ie02d2e1441ca2b3c7c159ffa917f7040f50327b5
Reviewed-on: http://review.typo3.org/34818
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>